requirements.txt
kafka-python>=2.0.2
streamlit>=1.20.0
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
plotly>=5.0.0